+++ dtoolcore/storagebroker.py
@@ -325,6 +325,16 @@ class DiskStorageBroker(object):
 
         return metadata
 
+    def pre_freeze_hook(self):
+        """Pre :meth:`dtoolcore.ProtoDataSet.freeze` actions.
+
+        This method is called at the beginning of the
+        :meth:`dtoolcore.ProtoDataSet.freeze` method.
+
+        It may be useful for remote storage backends to generate
+        caches to remove repetitive time consuming calls
+        """
+
     def post_freeze_hook(self):
         """Post :meth:`dtoolcore.ProtoDataSet.freeze` cleanup actions.